requirements.txt
python-socketio[client]
aiohttp
